Initial attempt at a batch communication.
It doesn't work entirely properly because although we check that all
parties are ready to communicate, we do not sync precisely when the
batch starts. Instead, we just sync with a blocking read, which is not
precise enough. It can't be precise enough because the poll takes 5
cycles, which is as large as the interval between flits.

It does "kind of" work though, which is progress.