Get broadcast to work from each core in turn.
The interleaving of the prints looks funny because the print from one of
the cores appears to happen a couple cycles later than the prints from
the other two cores. However, all 126 numbers are correctly received, in
the correct order, by all receiving cores. That's 126 numbers
broadcasted in roughly 1233 cycles. 126*5=630 of those cycles are
necessarily required by the TDM schedule, 115*4=460 of those cycles are
synchronization overhead at the start of sending a packet, and the
remaining 143 or so cycles are probably the C glue code.

Obviously the proportion that is overhead will vary depending on the
packet size -- smaller packets -> more packets for the same amount of
data -> more overhead. In this case 4 packets were sent.

By my measurement, the synchronization overhead incurred before sending
a packet -- and this is the assembly, not the C which is probably doing
a function call, saving things on the stack, etc. -- is 115 cycles.
Multiple measurements all gave something like 114.5 cycles.