Start extending the protocol.
I want to have multiple bursts, to amortize the initial hundred-cycle
synchronization overhead while still allowing enough time between bursts
for the sender to either jump back to the start of an unrolled loop or
for the sender to load words from main memory before sending them from
the register file.

In this commit the program in low_level_interface_noc still seems to
work.